refactor for located textbox handler logic
