+++ traitsui/testing/tester/qt4/helpers.py
@@ -128,56 +128,3 @@ def key_click_qwidget(control, interaction, delay):
     if not control.isEnabled():
         raise Disabled("{!r} is disabled.".format(control))
     key_click(control, interaction.key, delay=delay)
-
-
-def get_qobject_registry():
-     """ Creates a generic registry for handling/solving qt objects. (i.e.
-     this registry is independent of TraitsUI)
-
-     Returns
-     -------
-     registry : TargetRegistry
-         Registry containing qt specific generic handlers and solvers.
-     """
-     registry = TargetRegistry()
-
-     widget_classes = [
-         QtGui.QLineEdit,
-         QtGui.QTextEdit,
-         QtGui.QPushButton,
-     ]
-     handlers = [
-         (command.KeySequence, (lambda wrapper, interaction: key_sequence_qwidget(
-                                wrapper.target, interaction.sequence, wrapper.delay))),
-         (command.KeyClick, (lambda wrapper, interaction: key_click_qwidget(
-                             wrapper.target, interaction.key, wrapper.delay))),
-         (command.MouseClick, (lambda wrapper, _: mouse_click_qwidget(
-             wrapper.target, wrapper.delay))),
-     ]
-     for widget_class in widget_classes:
-         for interaction_class, handler in handlers:
-             registry.register_handler(
-                 target_class=widget_class,
-                 interaction_class=interaction_class,
-                 handler=handler,
-             )
-
-     registry.register_handler(
-         target_class=QtGui.QLineEdit,
-         interaction_class=query.DisplayedText,
-         handler=lambda wrapper, _: wrapper.target.displayText(),
-     )
-
-     registry.register_handler(
-         target_class=QtGui.QTextEdit,
-         interaction_class=query.DisplayedText,
-         handler=lambda wrapper, _: wrapper.target.toPlainText(),
-     )
-
-     registry.register_handler(
-         target_class=QtGui.QPushButton,
-         interaction_class=query.DisplayedText,
-         handler=lambda wrapper, _: wrapper.target.text(),
-     )
-
-     return registry
