+++ traitsui/testing/tester/qt4/implementation/range_editor.py
@@ -17,31 +17,12 @@ from traitsui.qt4.range_editor import (
     SimpleSpinEditor,
 )
 
-from traitsui.testing.tester import command, locator, query
-from traitsui.testing.tester.qt4 import helpers
+from traitsui.testing.tester import locator
+from traitsui.testing.tester.qt4.located_object_handlers import LocatedTextbox
 
 
-class RangeEditorTextbox:
-    def __init__(self, textbox):
-        self.textbox = textbox
-
-    @classmethod
-    def register(cls, registry):
-        handlers = [
-            (command.KeySequence, (lambda wrapper, interaction: helpers.key_sequence_qwidget(
-                                    wrapper.target.textbox, interaction.sequence, wrapper.delay))),
-            (command.KeyClick, (lambda wrapper, interaction: helpers.key_click_qwidget(
-                                wrapper.target.textbox, interaction.key, wrapper.delay))),
-            (command.MouseClick, (lambda wrapper, _: helpers.mouse_click_qwidget(
-                wrapper.target.textbox, wrapper.delay))),
-            (query.DisplayedText, lambda wrapper, _: wrapper.target.textbox.displayText()),
-        ]
-        for interaction_class, handler in handlers:
-            registry.register_handler(
-                target_class=cls,
-                interaction_class=interaction_class,
-                handler=handler,
-            )
+class RangeEditorTextbox(LocatedTextbox):
+    pass
 
 
 def resolve_location_simple_slider(wrapper, location):
