+++ traitsui/testing/tester/wx/default_registry.py
@@ -9,7 +9,7 @@
 #  Thanks for using Enthought open source!
 #
 
-from traitsui.testing.tester.wx.helpers import get_wobject_registry
+from traitsui.testing.tester.registry import TargetRegistry
 from traitsui.testing.tester.wx.implementation import (
     button_editor,
     range_editor,
@@ -26,7 +26,7 @@ def get_default_registry():
         The default registry containing implementations for TraitsUI editors
         that is wx specific.
     """
-    registry = get_wobject_registry()
+    registry = TargetRegistry()
 
     # ButtonEditor
     button_editor.register(registry)
