+++ traitsui/testing/tester/wx/helpers.py
@@ -137,47 +137,3 @@ def key_sequence_text_ctrl(control, interaction, delay):
         control.SetFocus()
     for char in interaction.sequence:
         key_click(control, char, delay)
-
-
-def get_wobject_registry():
-    """ Creates a generic registry for handling/solving wx objects. (i.e.
-    this registry is independent of TraitsUI)
-    Returns
-    -------
-    registry : TargetRegistry
-        Registry containing wx specific generic handlers and solvers.
-    """
-    registry = TargetRegistry()
-
-
-    handlers = [
-        (command.KeyClick, (lambda wrapper, interaction: key_click_text_ctrl(
-             wrapper.target, interaction.key, wrapper.delay))),
-        (command.KeySequence, (lambda wrapper, interaction: key_sequence_text_ctrl(
-             wrapper.target, interaction.sequence, wrapper.delay))),
-        (query.DisplayedText, lambda wrapper, _: wrapper.target.GetValue()),
-    ]
-
-    for interaction_class, handler in handlers:
-        registry.register_handler(
-            target_class=wx.TextCtrl,
-            interaction_class=interaction_class,
-            handler=handler
-        )
-
-    registry.register_handler(
-        target_class=wx.StaticText,
-        interaction_class=query.DisplayedText,
-        handler=lambda wrapper, action: (
-            wrapper.target.GetLabel()
-        ),
-    )
-
-    registry.register_handler(
-        target_class=wx.Button,
-        interaction_class=command.MouseClick,
-        handler=lambda wrapper, _: mouse_click_button(wrapper.target, wrapper.delay)
-    )
-
-    return registry
-
